static analysis fixes
